add operations for DB methods
